refactor: optz AVRecorder code
