+++ packages/av-recorder/package.json
@@ -37,7 +37,6 @@
     "@types/dom-webcodecs": "^0.1.7",
     "@types/wicg-file-system-access": "^2020.9.6",
     "@webav/av-cliper": "^0.9.0-beta.16",
-    "@webav/mp4box.js": "0.5.3-fenghen",
     "fix-webm-duration": "^1.0.5"
   }
 }
