--- packages/av-recorder/src/recorde.ts
@@ -1,172 +1,0 @@
-import mp4box from '@webav/mp4box.js';
-import { autoReadStream, file2stream, recodemux } from '@webav/av-cliper';
-import { TClearFn, IRecordeOpts } from './types';
-
-if (import.meta.env.DEV) {
-  mp4box.Log.setLogLevel(mp4box.Log.debug);
-}
-
-export class RecoderPauseCtrl {
-  // 当前帧的偏移时间，用于计算帧的 timestamp
-  #offsetTime = performance.now();
-
-  // 编码上一帧的时间，用于计算出当前帧的持续时长
-  #lastTime = this.#offsetTime;
-
-  // 用于限制 帧率
-  #frameCnt = 0;
-
-  // 如果为true，则暂停编码数据
-  // 取消暂停时，需要减去
-  #paused = false;
-
-  // 触发暂停的时间，用于计算暂停持续了多久
-  #pauseTime = 0;
-
-  constructor(readonly expectFPS: number) {}
-
-  start() {
-    this.#offsetTime = performance.now();
-    this.#lastTime = this.#offsetTime;
-  }
-
-  play() {
-    if (!this.#paused) return;
-    this.#paused = false;
-
-    this.#offsetTime += performance.now() - this.#pauseTime;
-    this.#lastTime += performance.now() - this.#pauseTime;
-  }
-
-  pause() {
-    if (this.#paused) return;
-    this.#paused = true;
-    this.#pauseTime = performance.now();
-  }
-
-  transfromVideo(frame: VideoFrame) {
-    const now = performance.now();
-    const offsetTime = now - this.#offsetTime;
-    if (
-      this.#paused ||
-      // 避免帧率超出期望太高
-      (this.#frameCnt / offsetTime) * 1000 > this.expectFPS
-    ) {
-      frame.close();
-      return;
-    }
-
-    const vf = new VideoFrame(frame, {
-      // timestamp 单位 微秒
-      timestamp: offsetTime * 1000,
-      duration: (now - this.#lastTime) * 1000,
-    });
-    this.#lastTime = now;
-
-    this.#frameCnt += 1;
-    frame.close();
-    return {
-      vf,
-      opts: { keyFrame: this.#frameCnt % 30 === 0 },
-    };
-  }
-
-  transformAudio(ad: AudioData) {
-    if (this.#paused) {
-      ad.close();
-      return;
-    }
-    return ad;
-  }
-}
-
-export function startRecorde(
-  opts: IRecordeOpts,
-  ctrl: RecoderPauseCtrl,
-  onEnded: TClearFn,
-) {
-  let stopEncodeVideo: TClearFn | null = null;
-  let stopEncodeAudio: TClearFn | null = null;
-
-  const recoder = recodemux({
-    video: { ...opts.video, bitrate: opts.bitrate ?? 3_000_000 },
-    audio: opts.audio,
-  });
-
-  let stoped = false;
-  if (opts.streams.video != null) {
-    let lastVf: VideoFrame | null = null;
-    let autoInsertVFTimer = 0;
-    const emitVf = (vf: VideoFrame) => {
-      clearTimeout(autoInsertVFTimer);
-
-      lastVf?.close();
-      lastVf = vf;
-      const vfWrap = ctrl.transfromVideo(vf.clone());
-      if (vfWrap == null) return;
-      recoder.encodeVideo(vfWrap.vf, vfWrap.opts);
-
-      // 录制静态画面，MediaStream 不出帧时，每秒插入一帧
-      autoInsertVFTimer = self.setTimeout(() => {
-        if (lastVf == null) return;
-        const newVf = new VideoFrame(lastVf, {
-          timestamp: lastVf.timestamp + 1e6,
-          duration: 1e6,
-        });
-        emitVf(newVf);
-      }, 1000);
-    };
-
-    ctrl.start();
-    const stopReadStream = autoReadStream(opts.streams.video, {
-      onChunk: async (chunk: VideoFrame) => {
-        if (stoped) {
-          chunk.close();
-          return;
-        }
-        emitVf(chunk);
-      },
-      onDone: () => {},
-    });
-
-    stopEncodeVideo = () => {
-      stopReadStream();
-      clearTimeout(autoInsertVFTimer);
-      lastVf?.close();
-    };
-  }
-
-  if (opts.audio != null && opts.streams.audio != null) {
-    stopEncodeAudio = autoReadStream(opts.streams.audio, {
-      onChunk: async (ad: AudioData) => {
-        if (stoped) {
-          ad.close();
-          return;
-        }
-        const newAD = ctrl.transformAudio(ad);
-        if (newAD != null) recoder.encodeAudio(ad);
-      },
-      onDone: () => {},
-    });
-  }
-
-  const { stream, stop: stopStream } = file2stream(
-    recoder.mp4file,
-    opts.timeSlice,
-    () => {
-      exit();
-      onEnded();
-    },
-  );
-
-  function exit() {
-    stoped = true;
-
-    stopEncodeVideo?.();
-    stopEncodeAudio?.();
-    recoder.close();
-    stopStream();
-  }
-
-  return { exit, stream };
-}
